Prevent undeploy failure if challenge does not have service
Some problems may not have an accompanying service and thus will
fail when the `remove_instance()` function attempts to remove it
from `/etc/xinetd.d`. Skip this step is `service == None`.

Sidenote - if problems should have some sort of service there, then
there's an issue where that doesn't happen.

+++ picoCTF-shell/hacksport/deploy.py
@@ -956,8 +956,9 @@ def remove_instances(path, instance_list):
             deployment_json_path = join(deployment_json_dir,
                                         "{}.json".format(instance_number))
 
-            logger.debug("...Removing xinetd service '%s'.", service)
-            os.remove(join(XINETD_SERVICE_PATH, service))
+            if service != None:
+                logger.debug("...Removing xinetd service '%s'.", service)
+                os.remove(join(XINETD_SERVICE_PATH, service))
 
             logger.debug("...Removing deployment directory '%s'.", directory)
             shutil.rmtree(directory)
